+++ src/std/wi_math.c
@@ -15,37 +15,37 @@ _math_single_arg_function(struct wi_state* state, double (*fn)(double x)) {
 }
 
 static void
-_math_abs(struct wi_state* state, int arg_count) {
+_math_abs(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, fabs);
 }
 
 static void
-_math_acos(struct wi_state* state, int arg_count) {
+_math_acos(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, acos);
 }
 
 static void
-_math_asin(struct wi_state* state, int arg_count) {
+_math_asin(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, asin);
 }
 
 static void
-_math_atan(struct wi_state* state, int arg_count) {
+_math_atan(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, atan);
 }
 
 static void
-_math_ceil(struct wi_state* state, int arg_count) {
+_math_ceil(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, ceil);
 }
 
 static void
-_math_clamp(struct wi_state* state, int arg_count) {
+_math_clamp(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real real = wi_arg_real(state, 1);
     wi_real min  = wi_arg_real(state, 2);
@@ -54,32 +54,32 @@ _math_clamp(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_cos(struct wi_state* state, int arg_count) {
+_math_cos(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, cos);
 }
 
 static void
-_math_deg(struct wi_state* state, int arg_count) {
+_math_deg(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real rad = wi_arg_real(state, 1);
     wi_push_real(state, rad * (180.0 / M_PI));
 }
 
 static void
-_math_exp(struct wi_state* state, int arg_count) {
+_math_exp(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, exp);
 }
 
 static void
-_math_floor(struct wi_state* state, int arg_count) {
+_math_floor(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, floor);
 }
 
 static void
-_math_mod(struct wi_state* state, int arg_count) {
+_math_mod(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real a = wi_arg_real(state, 1);
     wi_real b = wi_arg_real(state, 2);
@@ -87,7 +87,7 @@ _math_mod(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_log(struct wi_state* state, int arg_count) {
+_math_log(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real value = wi_arg_real(state, 1);
     wi_real base  = wi_arg_real(state, 2);
@@ -95,19 +95,19 @@ _math_log(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_log10(struct wi_state* state, int arg_count) {
+_math_log10(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, log10);
 }
 
 static void
-_math_ln(struct wi_state* state, int arg_count) {
+_math_ln(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, log);
 }
 
 static void
-_math_max(struct wi_state* state, int arg_count) {
+_math_max(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real max = wi_arg_real(state, 1);
 
@@ -123,7 +123,7 @@ _math_max(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_min(struct wi_state* state, int arg_count) {
+_math_min(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real min = wi_arg_real(state, 1);
 
@@ -139,7 +139,7 @@ _math_min(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_pow(struct wi_state* state, int arg_count) {
+_math_pow(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real base = wi_arg_real(state, 1);
     wi_real exp  = wi_arg_real(state, 2);
@@ -147,45 +147,45 @@ _math_pow(struct wi_state* state, int arg_count) {
 }
 
 static void
-_math_rad(struct wi_state* state, int arg_count) {
+_math_rad(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real deg = wi_arg_real(state, 1);
     wi_push_real(state, deg * (M_PI / 180.0));
 }
 
 static void
-_math_random(struct wi_state* state, int arg_count) {
+_math_random(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_push_real(state, (wi_real)rand() / ((wi_real)RAND_MAX + 1.0));
 }
 
 static void
-_math_round(struct wi_state* state, int arg_count) {
+_math_round(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, round);
 }
 
 static void
-_math_sign(struct wi_state* state, int arg_count) {
+_math_sign(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     wi_real real = wi_arg_real(state, 1);
     wi_push_real(state, real > 0 ? 1 : real < 0 ? -1 : 0);
 }
 
 static void
-_math_sin(struct wi_state* state, int arg_count) {
+_math_sin(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, sin);
 }
 
 static void
-_math_sqrt(struct wi_state* state, int arg_count) {
+_math_sqrt(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, sqrt);
 }
 
 static void
-_math_tan(struct wi_state* state, int arg_count) {
+_math_tan(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     _math_single_arg_function(state, tan);
 }
