+++ src/stm/wi_map.c
@@ -4,7 +4,7 @@
 #include "../core/wi_state.h"
 
 static struct wi_map*
-_check_arg_map(struct wi_state* state, int arg) {
+_check_arg_map(struct wi_state* state, uint8_t arg) {
     if (!wi_value_is_map(state->ffi_stack[arg])) {
         wi_state_error(state, "bad argument %i - expected a value of type map but got %s", arg,
                        wi_value_type(state->ffi_stack[arg]));
@@ -19,7 +19,7 @@ _check_arg1_map(struct wi_state* state) {
 }
 
 static void
-_map_copy(struct wi_state* state, int arg_count) {
+_map_copy(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* src  = _check_arg1_map(state);
     struct wi_map* dest = wi_new_map(state->gc);
@@ -28,7 +28,7 @@ _map_copy(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_clear(struct wi_state* state, int arg_count) {
+_map_clear(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map = _check_arg1_map(state);
     wi_table_free(&map->items);
@@ -36,21 +36,21 @@ _map_clear(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_capacity(struct wi_state* state, int arg_count) {
+_map_capacity(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map = _check_arg1_map(state);
     wi_push_real(state, map->items.capacity);
 }
 
 static void
-_map_count(struct wi_state* state, int arg_count) {
+_map_count(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map = _check_arg1_map(state);
     wi_push_real(state, map->items.live_count);
 }
 
 static void
-_map_keys(struct wi_state* state, int arg_count) {
+_map_keys(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map*   map    = _check_arg1_map(state);
     struct wi_array* result = wi_new_array(state->gc);
@@ -67,7 +67,7 @@ _map_keys(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_values(struct wi_state* state, int arg_count) {
+_map_values(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map*   map    = _check_arg1_map(state);
     struct wi_array* result = wi_new_array(state->gc);
@@ -85,7 +85,7 @@ _map_values(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_has(struct wi_state* state, int arg_count) {
+_map_has(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map    = _check_arg1_map(state);
     bool           exists = wi_table_get(&map->items, state->ffi_stack[2], NULL);
@@ -93,7 +93,7 @@ _map_has(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_get_or_default(struct wi_state* state, int arg_count) {
+_map_get_or_default(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map = _check_arg1_map(state);
     wi_value       value;
@@ -107,14 +107,14 @@ _map_get_or_default(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_remove(struct wi_state* state, int arg_count) {
+_map_remove(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map = _check_arg1_map(state);
     wi_push_bool(state, wi_table_delete(&map->items, state->ffi_stack[2]));
 }
 
 static void
-_map_each(struct wi_state* state, int arg_count) {
+_map_each(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map       = _check_arg1_map(state);
     int            mod_count = map->items.mod_count;
@@ -139,7 +139,7 @@ _map_each(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_select(struct wi_state* state, int arg_count) {
+_map_select(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map       = _check_arg1_map(state);
     int            mod_count = map->items.mod_count;
@@ -177,7 +177,7 @@ _map_select(struct wi_state* state, int arg_count) {
 }
 
 static void
-_map_where(struct wi_state* state, int arg_count) {
+_map_where(struct wi_state* state, uint8_t arg_count) {
     WI_UNUSED(arg_count);
     struct wi_map* map       = _check_arg1_map(state);
     int            mod_count = map->items.mod_count;
